[+] fix compilation errors for parser and rebase the architecture of parser part
